merge ian to master

+++ album/views.py
@@ -8,10 +8,8 @@ from django.views.decorators.csrf import csrf_exempt
 @csrf_exempt
 def album_create(request):
     if request.method == 'POST':
-        judul = request.POST.get('judul')
+        judul = request.POST.get('judul_album')
         id_label = request.POST.get('label')
-        print(judul)
-        print(id_label)
         id_album = str(uuid.uuid4())
         query_str = f"""INSERT INTO album (id, judul, jumlah_lagu, id_label, total_durasi)
             VALUES (
@@ -23,15 +21,12 @@ def album_create(request):
             )
             """
         res = query(query_str)
-        print(res)
 
         judul = request.POST.get('judul')
         artist_email = request.POST.get('artist_email')
         songwriters_email = request.POST.getlist('songwriter')
         genres = request.POST.getlist('genre')
         durasi = request.POST.get('durasi')
-        
-        print(artist_email)
 
         id_konten = str(uuid.uuid4())
         current_date = datetime.now()
@@ -91,6 +86,16 @@ def album_create(request):
 
         #Ga perlu update durasi sm jumlah lagu di album gausah soalnya pake trigger
 
+        #Tambahin royalti
+        list_pemilik_hak_cipta = []
+        list_pemilik_hak_cipta += query(f"SELECT id_pemilik_hak_cipta FROM artist WHERE id = '{id_artist}'")
+        list_pemilik_hak_cipta += query(f"SELECT id_pemilik_hak_cipta FROM songwriter WHERE id in {tuple(songwriters_id)}")
+        list_pemilik_hak_cipta += query(f"SELECT l.id_pemilik_hak_cipta FROM label l JOIN album a ON a.id_label = l.id WHERE a.id = '{id_album}'")
+        ids_pemilik_hak_cipta = [pemilik['id_pemilik_hak_cipta'] for pemilik in list_pemilik_hak_cipta]
+
+        for ids in ids_pemilik_hak_cipta:
+            print(query(f"INSERT INTO royalti VALUES('{ids}', '{id_konten}', 0)"))
+
         return redirect('album:album_list')
 
     query_str = """SELECT a.nama, a.email
@@ -115,8 +120,7 @@ def album_create(request):
     query_str = "SELECT nama, id FROM LABEL"
     labels = query(query_str)
 
-
-    return render(request, 'album_form.html', {
+    data = {
         'labels': labels,
         'role': akun['role'],
         'name': nama_akun,
@@ -124,30 +128,68 @@ def album_create(request):
         'artists': artists,
         'songwriters': songwriters,
         'genres': genres
-    })
-
-def album_list(request):
-    query_str = """SELECT a.id, a.judul, l.nama as label, a.jumlah_lagu, a.total_durasi
-    FROM album as a
-    JOIN label as l ON a.id_label = l.id
-    """
-    result = query(query_str)
-
-    akun = request.session.get('akun', None)
-    print(akun)
-    role = ''
-    role = "Label" if akun and not akun.get('role', None) else "Bukan Label"
-
-    data = {'albums': result, 'role': role}
+    }
     data.update(akun)
 
-    if role != "Label":
+    return render(request, 'album_form.html', data)
+
+def album_list(request):
+    if not request.session.get('akun', None):
+        return redirect('main:dashboard')
+    akun = request.session.get('akun')
+    role = akun.get('role', 'podcaster')
+    if role == 'podcaster':
+        return redirect('main:dashboard')
+
+    if role != "label":
+        query_str = f"""
+        SELECT A.judul, L.nama AS label, A.jumlah_lagu, A.total_durasi, A.id
+        FROM ALBUM A
+        JOIN LABEL L ON A.id_label = L.id
+        JOIN SONG S ON A.id = S.id_album
+        JOIN ARTIST Ar ON S.id_artist = Ar.id
+        JOIN AKUN Ak ON Ar.email_akun = Ak.email
+        WHERE Ak.email = '{akun['email']}'
+
+        UNION
+
+        SELECT A.judul, L.nama AS label, A.jumlah_lagu, A.total_durasi, A.id
+        FROM ALBUM A
+        JOIN LABEL L ON A.id_label = L.id
+        JOIN SONG S ON A.id = S.id_album
+        JOIN SONGWRITER_WRITE_SONG SW ON S.id_konten = SW.id_song
+        JOIN SONGWRITER So ON SW.id_songwriter = So.id
+        JOIN AKUN Ak ON So.email_akun = Ak.email
+        WHERE Ak.email = '{akun['email']}'
+
+        ORDER BY judul ASC;
+        """
+    else:
+        query_str = f"""SELECT A.judul, L.nama AS label, A.jumlah_lagu, A.total_durasi, A.id
+            FROM ALBUM A, LABEL L
+            WHERE A.id_label = L.id AND L.email = '{akun['email']}'
+            ORDER BY A.judul ASC"""
+    result = query(query_str)
+    print(result)
+
+    data = {'albums': result}
+    data.update(akun)
+
+    if role != "label":
         return render(request, 'album_list.html', data)
     else:
         return render(request, 'album_list_label.html', data)
 
 @csrf_exempt
 def tambah_lagu(request, id):
+    akun = request.session.get('akun', None)
+    if not akun:
+        return redirect('main:page_login')
+    role = akun.get('role', 'bukan_siapa_siapa')
+    print(role)
+    print(akun)
+    if not (role == 'artis' or role == 'songwriter'):
+        return redirect('main:dashboard')
     if request.method == 'POST':
         judul = request.POST.get('judul')
         artist_email = request.POST.get('artist_email')
@@ -199,7 +241,7 @@ def tambah_lagu(request, id):
 
         #Tambah songwriter_write_song utk semua songwriter
         songwriters = query("SELECT id, email_akun FROM songwriter")
-        songwriters_id = [songwriter['id'] for songwriter in songwriters if songwriter['email_akun'] in songwriters_email]
+        songwriters_id = [str(sw['id']) for sw in songwriters if sw['email_akun'] in songwriters_email]
 
         for songwriter_id in songwriters_id:
             query_str = f"""INSERT INTO songwriter_write_song (id_songwriter, id_song)
@@ -215,6 +257,17 @@ def tambah_lagu(request, id):
 
         #Ga perlu update durasi sm jumlah lagu di album gausah soalnya pake trigger
 
+        #Tambahin royalti
+        list_pemilik_hak_cipta = []
+        list_pemilik_hak_cipta += query(f"SELECT id_pemilik_hak_cipta FROM artist WHERE id = '{id_artist}'")
+        list_pemilik_hak_cipta += query(f"SELECT id_pemilik_hak_cipta FROM songwriter WHERE id in {tuple(songwriters_id)}")
+        list_pemilik_hak_cipta += query(f"SELECT l.id_pemilik_hak_cipta FROM label l JOIN album a ON a.id_label = l.id WHERE a.id = '{id}'")
+        print(list_pemilik_hak_cipta)
+        ids_pemilik_hak_cipta = [pemilik['id_pemilik_hak_cipta'] for pemilik in list_pemilik_hak_cipta]
+
+        for ids in ids_pemilik_hak_cipta:
+            print(query(f"INSERT INTO royalti VALUES('{ids}', '{id_konten}', 0)"))
+
         return redirect('album:album_list')
     elif 'role' not in request.session.get('akun', {}):
         return redirect('album:album_list')
@@ -222,7 +275,6 @@ def tambah_lagu(request, id):
         query_str = f"SELECT * from album WHERE id = '{id}'"
         album = query(query_str)[0]
         album['id'] = str(album['id'])
-        print(album)
 
         query_str = """SELECT a.nama, a.email
         FROM akun as a
@@ -242,7 +294,7 @@ def tambah_lagu(request, id):
         akun = request.session.get('akun', None)
         nama_akun = akun['nama'] if akun else ''
 
-        return render(request, 'lagu_add.html', {
+        data = {
             'album': album,
             'role': akun['role'],
             'name': nama_akun,
@@ -250,9 +302,21 @@ def tambah_lagu(request, id):
             'artists': artists,
             'songwriters': songwriters,
             'genres': genres
-        })
+        }
+        data.update(akun)
+
+        return render(request, 'lagu_add.html', data)
 
 def daftar_lagu(request, id):
+    akun = request.session.get('akun', None)
+    if not akun:
+        return redirect('main:page_login')
+    role = akun.get('role', 'podcaster')
+    if role == 'podcaster':
+        return redirect('main:dashboard')
+    
+    #Asumsi yang ditampilkan adalah seluruh lagu pada album yg terkait bukan hanya milik artis/songwriter/labelnya
+
     query_str = f"SELECT judul FROM album WHERE id = '{id}'"
     judul_album = query(query_str)
 
@@ -263,29 +327,56 @@ def daftar_lagu(request, id):
     """
     songs = query(query_str)
 
-    return render(request, 'lagu_list.html', {'album': judul_album, 'songs': songs})
+    data = {'album': judul_album, 'songs': songs}
+    data.update(request.session.get('akun', None))
+
+    return render(request, 'lagu_list.html', data)
 
 @csrf_exempt
 def delete_album(request, id):
+    akun = request.session.get('akun', None)
+    if not akun:
+        return redirect('main:page_login')
+    role = akun.get('role', 'podcaster')
+    if role == 'podcaster':
+        return redirect('main:dashboard')
+    
     if request.method == 'POST':
-        query(f"DELETE FROM album WHERE id = '{id}'")
+        songs = query(f"SELECT id_konten FROM song WHERE id_album = '{id}'")
+        print(songs)
+        for ids in songs:
+            delete_song(ids['id_konten'])
+
+        s = query(f"DELETE FROM album WHERE id = '{id}'")
+        print(s)
         return JsonResponse({'status': 'success'})
     return HttpResponseNotFound()
 
 @csrf_exempt
 def delete_lagu(request, id):
+    akun = request.session.get('akun', None)
+    if not akun:
+        return redirect('main:page_login')
+    role = akun.get('role', 'podcaster')
+    if role == 'podcaster':
+        return redirect('main:dashboard')
+
     if request.method == 'POST':
-        status = True
-        res = query(f"DELETE FROM genre WHERE id_konten = '{id}'")
-        status = status and isinstance(res, int)
-        res = query(f"DELETE FROM songwriter_write_song WHERE id_song = '{id}'")
-        status = status and isinstance(res, int)
-        res = query(f"DELETE FROM song WHERE id_konten = '{id}'")
-        status = status and isinstance(res, int)
-        res = query(f"DELETE FROM konten WHERE id = '{id}'")
-        status = status and isinstance(res, int)
-        if res:
+        status = delete_song(id)
+        if status:
             return JsonResponse({'status': 'success'})
         else:
             return JsonResponse({'status': 'failed'})
     return HttpResponseNotFound()
+
+def delete_song(id):
+    status = True
+    res = query(f"DELETE FROM genre WHERE id_konten = '{id}'")
+    status = status and isinstance(res, int)
+    res = query(f"DELETE FROM songwriter_write_song WHERE id_song = '{id}'")
+    status = status and isinstance(res, int)
+    res = query(f"DELETE FROM song WHERE id_konten = '{id}'")
+    status = status and isinstance(res, int)
+    res = query(f"DELETE FROM konten WHERE id = '{id}'")
+    status = status and isinstance(res, int)
+    return status
